% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ims_article.R
\name{ims_article}
\alias{ims_article}
\title{Elsevier journal format.}
\usage{
ims_article(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"))
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}}
for additional details.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Elsevier journal format.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "ims_article", package = "imsart")
}

}

